docs: fix lint issues on main
main is currently red due to lint failures. Ran super-linter locally and
see the lots of errors. This commits attempts to fix all the lint
errors.

A handful of the errors were links to other sections of the same
document in the Chinese version of the files. Both ebpf_in_kepler.md and
ebpf_in_kepler.zh.md had a `Contents` section at the top with a link to
all the other sections. This really isn't needed since the mkdocs
generated website has a Table of Contents built in. By removing this,
that fix a lot of the link errors. The remaining few requires assigning
a name to the section being linked to (in kepler-operator.zh.md).

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>